move workflow creation out of distribution module, and isolate dropwizard components
